% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{activity}
\description{
xAPI Activity object definitions
}
\section{Details}{

 A Statement can represent an Activity as the Object of the Statement.
}

\seealso{
\link{https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#activity-definition}
}
