% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor.R
\name{actor}
\alias{actor}
\title{actor}
\description{
xAPI Actor object definitions
}
\section{Description}{

  The Actor defines who performed the action. The Actor of a Statement can be an Agent or a Group.
}

\seealso{
\link{https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#actor}
}
