% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{api_request}
\alias{api_request}
\title{Makes a simple API request to Learning Locker endpoint
Uses HTTP Connection Interface}
\usage{
api_request(session, config, request, asJSON = FALSE)
}
\arguments{
\item{session}{The current R session}

\item{config}{Endpoint configuration}

\item{request}{API request string}

\item{asJSON}{(optional) Return content as json string}
}
\value{
response content
}
\description{
Makes a simple API request to Learning Locker endpoint
Uses HTTP Connection Interface
}
\details{
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html
}
\seealso{
\link{https://docs.learninglocker.net/http-connection/}
}
