% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{checkSupportedComponents}
\alias{checkSupportedComponents}
\title{checkSupportedComponents}
\usage{
checkSupportedComponents(object)
}
\description{
Check to see if provided Interaction Type supports the given Component.
Not all types will support all components as per the xAPI Specification.
}
