% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Sets up scripts needed for Learning Locker API}
\usage{
connect(session, config)
}
\arguments{
\item{session}{The current R session}

\item{config}{API Configuration}
}
\value{
HTTP Status Code
}
\description{
Sets up scripts needed for Learning Locker API
}
