% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{context}
\description{
xAPI Context object definitions
}
\section{Details}{

  An optional property that provides a place to add contextual information to a Statement.
}

\seealso{
\link{https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#context}
}
