% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor.R
\name{createActor}
\alias{createActor}
\title{createActor}
\usage{
createActor(agent = NULL, group = NULL, warn = FALSE, ...)
}
\arguments{
\item{agent}{Agent params}

\item{group}{Group params}

\item{warn}{Show warnings}
}
\value{
Actor
}
\description{
createActor
}
