% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{createContext}
\alias{createContext}
\title{createContext}
\usage{
createContext(context = NULL, warn = FALSE, ...)
}
\arguments{
\item{context}{Context params}
}
\value{
Context
}
\description{
createContext
}
