% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{createExtension}
\alias{createExtension}
\title{Creates an xAPI Extension object}
\usage{
createExtension(extension = NULL, warn = FALSE, ...)
}
\arguments{
\item{ref}{Internationalized resource identifier (uri extension)}

\item{value}{Any value matching corresponding reference definition constraints}
}
\value{
xAPI Extension object
}
\description{
Creates an xAPI Extension object
}
\examples{
createExtension(extension = list(ref = "https://w3id.org/xapi/cmi5/result/extensions/progress", value = 100))

}
\seealso{
\code{extension}
}
