% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{createGroup}
\alias{createGroup}
\title{Creates an xAPI Group object}
\usage{
createGroup(group = NULL, warn = FALSE, ...)
}
\arguments{
\item{group}{Group params}

\item{warn}{Show warnings}
}
\value{
xAPI Group object
}
\description{
Creates an xAPI Group object
}
\examples{
createGroup(
  group = list(
    name = "RTeam",
    members = list(
      agent(name = "Bob", mbox = "mailto:bob@example.com"),
      agent(name = "Kathy", mbox = "mailto:kathy@example.com"),
      agent(name = "Tom", mbox = "mailto:tom@example.com")
    )
  )
)

}
\seealso{
\code{\link{group}}
}
