% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{createObject}
\alias{createObject}
\title{Creates an xAPI Object object}
\usage{
createObject(object = NULL, warn = FALSE, ...)
}
\arguments{
\item{object}{Object params}

\item{warn}{Show warnings}
}
\value{
xAPI Object object
}
\description{
Creates an xAPI Object object
}
\examples{
createObject(object = list(name = "Question 1", description = "Example question description."))

}
\seealso{
\code{object}
}
