% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createResult}
\alias{createResult}
\title{Creates an xAPI Result object}
\usage{
createResult(result = NULL, warn = FALSE, ...)
}
\arguments{
\item{warn}{Show warnings}

\item{success}{Indicates whether or not the attempt on the Activity was successful.}

\item{response}{A response appropriately formatted for the given Activity.}
}
\value{
xAPI Result object
}
\description{
Creates an xAPI Result object
}
\section{Extras}{

This function accepts an optional result value \code{extension}.
}

\examples{
createResult(result = list(success = TRUE, response = "correctAnswer"))

}
\seealso{
\code{\link{result}}

\code{\link{createExtension}}
}
