% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{createScore}
\alias{createScore}
\title{Creates an xAPI Score object}
\usage{
createScore(score = NULL, warn = FALSE, ...)
}
\arguments{
\item{score}{Score params}

\item{warn}{Show warnings; ignored for important warnings.}
}
\value{
xAPI Score object
}
\description{
Creates an xAPI Score object
}
\examples{
createScore(score = list(scaled = 0, raw = 80, min = 0, max = 100))

}
\seealso{
\code{\link{score}}
}
