% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createStatement}
\alias{createStatement}
\title{Creates an xAPI Statement}
\usage{
createStatement(x = NULL, warn = TRUE, ...)
}
\arguments{
\item{x}{Statement values}

\item{warn}{Show warnings}
}
\value{
xAPI Statement (json)
}
\description{
Creates an xAPI Statement
}
\examples{
createStatement(success = TRUE, response = "correctAnswer")

}
\seealso{
\code{\link{createActor}}

\code{\link{createVerb}}

\code{\link{createObject}}

\code{\link{createResult}}
}
