% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createVerb}
\alias{createVerb}
\title{Creates an xAPI Verb object}
\usage{
createVerb(verb = NULL, id = NULL, warn = TRUE, ...)
}
\arguments{
\item{verb}{Verb name}

\item{id}{URI to verb definition}

\item{warn}{Show warnings}
}
\value{
xAPI Verb object
}
\description{
Creates an xAPI Verb object
}
\examples{
createVerb(verb = "experienced")
createVerb(verb = "custom-verb", id = "https://example.com/xapi/verbs/custom-verb")

}
\seealso{
\code{\link{verbs}}
}
