% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{formatJSON}
\alias{formatJSON}
\title{formatJSON}
\usage{
formatJSON(json, ...)
}
\arguments{
\item{json}{JSON}

\item{...}{jsonlite args}
}
\value{
json
}
\description{
Uses jsonlite to format json with a set of default params.
}
