% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_system_language_tag}
\alias{get_system_language_tag}
\title{Returns System RFC 5646 Language Tag}
\usage{
get_system_language_tag()
}
\value{
Always returns en-US
}
\description{
Returns System RFC 5646 Language Tag
}
