% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{getContextDefinition}
\alias{getContextDefinition}
\title{getContextDefinition}
\usage{
getContextDefinition()
}
\value{
definition
}
\description{
Returns an empty Context object with possible arguments.
}
