% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{getExtensionDefinition}
\alias{getExtensionDefinition}
\title{getExtensionDefinition}
\usage{
getExtensionDefinition()
}
\value{
definition
}
\description{
Returns an empty Extension object with possible arguments.
}
