% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{getInteractionComponent}
\alias{getInteractionComponent}
\title{getInteractionComponent}
\usage{
getInteractionComponent(name, asJSON = FALSE)
}
\value{
Component
}
\description{
Returns details about a specific Interaction Component.
}
