% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{getInteractionComponents}
\alias{getInteractionComponents}
\title{getInteractionComponents}
\usage{
getInteractionComponents()
}
\value{
vector
}
\description{
Returns a list of the default Interaction Components.
}
