% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{getInteractionType}
\alias{getInteractionType}
\title{getInteractionType}
\usage{
getInteractionType(name, asJSON = FALSE)
}
\value{
Interaction
}
\description{
Returns details about a specific Interaction type.
}
