% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{getObjectDefinition}
\alias{getObjectDefinition}
\title{object}
\usage{
getObjectDefinition()
}
\description{
xAPI Object object definitions
}
\section{Description}{

  The Object defines the thing that was acted on. The Object of a Statement can be an Activity, Agent/Group, SubStatement, or Statement Reference.
}

\section{Details}{

  Objects which are provided as a value for this property SHOULD include an "objectType" property. If not specified, the objectType is assumed to be Activity. Other valid values are: Agent, Group, SubStatement or StatementRef. The properties of an Object change according to the objectType.
}

\seealso{
\link{https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#object}
}
