% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{getSupportedComponents}
\alias{getSupportedComponents}
\title{getSupportedComponents}
\usage{
getSupportedComponents(interactionType)
}
\value{

}
\description{
Returns a list of Components supported by a given Interaction type.
}
