% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb.R
\name{getVerb}
\alias{getVerb}
\title{Get Verb from xAPI Vocabulary}
\usage{
getVerb(verb = "", warn = FALSE, asJSON = FALSE)
}
\arguments{
\item{verb}{Name of verb}

\item{warn}{Show warnings}

\item{asJSON}{Returns object as JSON string}
}
\value{
Verb
}
\description{
Get Verb from xAPI Vocabulary
}
