% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{Interface}
\alias{Interface}
\alias{getInterfaceList}
\alias{getInterface}
\title{Interface}
\usage{
getInterfaceList()

getInterface(name, asJSON = FALSE)
}
\arguments{
\item{name}{Interface name}

\item{asJSON}{Return as json}
}
\value{
vector

Interface
}
\description{
Learning Locker API Interface

Returns details about a specific Learning Locker API Interface.
}
\seealso{
\link{https://docs.learninglocker.net/http-rest/}
getInterfaceList

Returns a vector of the Learning Locker API Interface types.
}
