% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachLocker.R
\name{lockerTest}
\alias{lockerTest}
\title{attachLocker
learningLocker}
\usage{
lockerTest(endpoint = "http://localhost:8000/xapi/", auth = NULL,
  user = NULL, password = NULL)
}
\arguments{
\item{endpoint}{Learning Locker API - Endpoint}

\item{auth}{Learning Locker API - Auth Token}

\item{user}{Learning Locker API - Auth User}

\item{password}{Learning Locker API - Auth Password}
}
\description{
attachLocker
learningLocker
}
