% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model}
\alias{Model}
\alias{getModelList}
\alias{getModel}
\title{Model}
\usage{
getModelList()

getModel(name, asJSON = FALSE)
}
\arguments{
\item{name}{Model name}

\item{asJSON}{Return as json}
}
\value{
vector

Model
}
\description{
Learning Locker API Model

Returns details of a specific Learning Locker API Model.
}
\seealso{
\link{https://docs.learninglocker.net/http-rest/}
getModelList

Returns a vector of Learning Locker API Models.
}
