% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\title{result}
\description{
xAPI Result object definitions
}
\section{Description}{

 An optional property that represents a measured outcome related to the Statement in which it is included.
}

\seealso{
\link{https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#result}
}
