% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlocker.R
\docType{package}
\name{rlocker}
\alias{rlocker}
\alias{rlocker-package}
\title{rlocker: Adds Learning Locker xAPI support for Shiny Applications.}
\description{
rlocker: Adds Learning Locker xAPI support for Shiny Applications.
}
\section{rlocker functions}{

The rlocker functions ...
}

