% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{rlockerJS}
\alias{rlockerJS}
\title{rlockerJS}
\format{
An object of class \code{html_dependency} of length 10.
}
\usage{
rlockerJS
}
\description{
htmlDependency js and css will be used in other functions with attachDependency
}
\keyword{datasets}
