% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{score}
\description{
xAPI Score object definitions
}
\section{Description}{

 An optional property that represents the outcome of a graded Activity achieved by an Agent.
}

\section{Requirements}{

- The Score Object SHOULD include "scaled" if a logical percent based score is known.
- The Score Object SHOULD NOT be used for scores relating to progress or completion. Consider using an extension (preferably from an established Community of Practice) instead.
}

\seealso{
\link{https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#2451-score}
}
