% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{store}
\alias{store}
\title{Stores an xAPI Statement in configured locker}
\usage{
store(session, statement = NULL, warn = FALSE, ...)
}
\arguments{
\item{session}{Shiny session}

\item{statement}{xAPI Statement}

\item{warn}{Show warnings}
}
\value{
HTTP Status
}
\description{
Stores an xAPI Statement in configured locker
}
\note{
Requires interactive session
}
\seealso{
\code{\link{connect}}

\code{\link{retrieve}}

\code{\link{createStatement}}
}
