% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{test}
\alias{test}
\title{Simple function to check if connection succeeds}
\usage{
test(config = get_locker_config())
}
\arguments{
\item{config}{Endpoint configuration}
}
\description{
Simple function to check if connection succeeds
}
\details{
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html
}
