% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb.R
\name{verb}
\alias{verb}
\title{verb}
\description{
xAPI Verb object definitions
}
\section{Description}{

  The Verb defines the action between an Actor and an Activity.
}

\section{Details}{

  Verbs appear in Statements as Objects consisting of an IRI and a set of display names corresponding to multiple languages or dialects which provide human-readable meanings of the Verb. The table below lists all properties of the Verb Object.
}

\seealso{
\link{https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#verbs}
}
