% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\docType{data}
\name{verbs}
\alias{verbs}
\title{verbs}
\format{An object of class \code{list} of length 28.}
\usage{
verbs
}
\description{
verbs
}
\details{
https://github.com/adlnet-archive/xAPIVerbs
}
\keyword{datasets}
