% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{xAPIWrapper}
\alias{xAPIWrapper}
\title{xAPIWrapper}
\format{
An object of class \code{html_dependency} of length 10.
}
\usage{
xAPIWrapper
}
\description{
htmlDependency js and css that will be used in other functions with attachDependency
}
\keyword{datasets}
