% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(session, config)
}
\arguments{
\item{session}{The current R session}

\item{config}{Endpoint configuration}
}
\value{
HTTP Status Code
}
\description{
connect

Sets up scripts needed for Learning Locker API
}
