% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createActor}
\alias{createActor}
\title{Creates an xAPI Actor object}
\usage{
createActor(actor = NULL, type = NULL, warn = TRUE, ...)
}
\arguments{
\item{actor}{Current user}

\item{type}{Type of actor}

\item{warn}{Show warnings}
}
\value{
xAPI Actor object
}
\description{
Creates an xAPI Actor object
}
\examples{
createActor(type = "Agent")

}
