% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createActor}
\alias{createActor}
\title{Creates an xAPI Actor object}
\usage{
createActor(actor = NULL, warn = TRUE, ...)
}
\arguments{
\item{actor}{Actor defaults}

\item{warn}{Show warnings}
}
\value{
xAPI Actor object
}
\description{
Creates an xAPI Actor object
}
\examples{
createActor(name = "John Doe", mbox = "mailto:john@example.com")

}
\seealso{
\code{\link{actor}}
}
