% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createAgent}
\alias{createAgent}
\title{Creates an xAPI Agent object}
\usage{
createAgent(agent = NULL, warn = FALSE, ...)
}
\arguments{
\item{agent}{Agent defaults}

\item{warn}{Show warnings}
}
\value{
xAPI Agent object
}
\description{
Creates an xAPI Agent object
}
\examples{
createAgent(agent = list(name = "John Doe", mbox = "mailto:john@example.com"))

}
\seealso{
\code{\link{agent}}
}
