% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createObject}
\alias{createObject}
\title{Creates an xAPI Object object}
\usage{
createObject(object = NULL, warn = TRUE, ...)
}
\arguments{
\item{warn}{Show warnings}

\item{name}{Object name}

\item{description}{Brief description of the object}

\item{id}{URI to object or activity}
}
\value{
xAPI Object object
}
\description{
Creates an xAPI Object object
}
\examples{
createObject(name = "Question 1", description = "Example question description.")
createObject(name = question$title, description = question$description, id = session$clientData)

}
