% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{createResult}
\alias{createResult}
\title{Creates an xAPI Result object}
\usage{
createResult(result = NULL, warn = FALSE, ...)
}
\arguments{
\item{result}{Result params}

\item{warn}{Show warnings}
}
\value{
Result
}
\description{
Creates an xAPI Result object
}
\section{Extras}{

This function accepts an optional result value \code{extension}.
}

\examples{
createResult(
  result = list(
    success = TRUE,
    response = "correctAnswer"
  )
)

createResult(
  result = list(
    success = TRUE,
    response = "correctAnswer",
    extensions = list(
      ref = "https://w3id.org/xapi/cmi5/result/extensions/progress", 
      value = 100
    )
  )
)

}
\seealso{
\code{\link{result}}

\code{\link{createExtension}}
}
