% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createResult}
\alias{createResult}
\title{Creates an xAPI Result object}
\usage{
createResult(success = NULL, response = NULL, warn = TRUE, ...)
}
\arguments{
\item{success}{Holds a value if interaction can be marked as successful/unsuccessful.}

\item{response}{Response from the user on a given action}

\item{warn}{Show warnings}
}
\value{
xAPI Result object
}
\description{
Creates an xAPI Result object
}
\examples{
createResult(success = TRUE, response = "correctAnswer")

}
