% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createStatement}
\alias{createStatement}
\title{Creates an xAPI Statement}
\usage{
createStatement(x = NULL, warn = TRUE, ...)
}
\arguments{
\item{x}{Statement values}

\item{warn}{Show warnings}
}
\value{
xAPI Statement (json)
}
\description{
Creates an xAPI Statement
}
\examples{
createStatement(
 list(
  agent = currentUser,
  verb = list(
  ),
  object = list(
   id = paste0(getCurrentAddress(session), "#", question$id),
   name = question$title,
   description = question$text
 ),
 result = list(
   success = session$input[[question$id]] == question$answer,
   response = session$input[[question$id]]
 )
)
)

}
\seealso{
\code{\link{createAgent}}

\code{\link{createVerb}}

\code{\link{createObject}}

\code{\link{createResult}}
}
