% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{createVerb}
\alias{createVerb}
\title{Creates an xAPI Verb object}
\usage{
createVerb(verb = NULL, warn = FALSE, ...)
}
\arguments{
\item{verb}{Verb name}

\item{warn}{Show warnings}

\item{id}{URI to verb definition}
}
\value{
xAPI Verb object
}
\description{
Creates an xAPI Verb object
}
\examples{
createVerb(verb = list(display = "custom-verb", id = "https://example.com/xapi/verbs/custom-verb"))

}
\seealso{
\code{\link{verbs}}
}
