% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verb.R
\name{createVerb}
\alias{createVerb}
\title{Creates an xAPI Verb object}
\usage{
createVerb(verb = NULL, warn = FALSE, asJSON = FALSE, ...)
}
\arguments{
\item{verb}{Verb params}

\item{warn}{Show warnings}

\item{asJSON}{Returns object as JSON string}
}
\value{
Verb
}
\description{
Creates an xAPI Verb object
}
\examples{
createVerb(
  verb = list(
    display = "custom-verb",
    id = "https://example.com/xapi/verbs/custom-verb"
  )
)

}
\seealso{
\code{verb}
}
