% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{Interface}
\alias{Interface}
\alias{getInterfaceList}
\alias{getInterface}
\title{Interface}
\usage{
getInterfaceList()

getInterface(name, asJSON = FALSE)
}
\description{
Learning Locker API Interface
}
\seealso{
\link{https://docs.learninglocker.net/http-rest/}
}
