% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model}
\alias{Model}
\alias{getModelList}
\alias{getModel}
\title{Model}
\usage{
getModelList()

getModel(name, asJSON = FALSE)
}
\description{
Learning Locker API Model
}
\seealso{
\link{https://docs.learninglocker.net/http-rest/}
}
