% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{retrieve}
\alias{retrieve}
\title{Makes a simple HTTP GET request to Learning Locker API endpoint
Uses HTTP Connection Interface}
\usage{
retrieve(model = NULL, interface = "connection", query,
  asJSON = FALSE)
}
\arguments{
\item{asJSON}{(optional) Return content as json string}

\item{session}{The current R session}

\item{request}{API request string}
}
\value{
response content
}
\description{
Makes a simple HTTP GET request to Learning Locker API endpoint
Uses HTTP Connection Interface
}
\details{
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html
}
\seealso{
\link{https://docs.learninglocker.net/http-statements/}

### ### ### ### ### ### ### ### ### THESE NEED TO BE UPDATED ### ### ### ### ### ### ### ### ###
Should this be renamed to retrieve as we're only making get requests? api_retrieve rlocker::retrieve
api/connection should be added to the base request automatically, we're not supporting REST at this time
base_url + "api/connection" + request_type + query
create new functions for connection details set_locker_config() / get_locker_config() (there's a collision with set_config in httr)
}
