% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{rLockerJS}
\alias{rLockerJS}
\title{rLockerJS}
\format{
An object of class \code{html_dependency} of length 10.
}
\usage{
rLockerJS
}
\description{
htmlDependency js and css will be used in other functions with attachDependency
}
\keyword{datasets}
