% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{store}
\alias{store}
\title{Stores an xAPI Statement}
\usage{
store(statement = NULL, warn = TRUE, ...)
}
\arguments{
\item{statement}{xAPI Statement}

\item{warn}{Show warnings}
}
\value{
HTTP Status
}
\description{
Stores an xAPI Statement
}
\examples{
store(statement, connection)

}
\seealso{
\code{\link{createStatement}}
}
