% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{store}
\alias{store}
\title{Stores an xAPI Statement}
\usage{
store(session, statement = NULL, warn = FALSE, ...)
}
\arguments{
\item{statement}{xAPI Statement}

\item{warn}{Show warnings}
}
\value{
HTTP Status
}
\description{
Stores an xAPI Statement
}
\seealso{
\code{\link{createStatement}}
}
