% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{test}
\alias{test}
\title{connect}
\usage{
test(config)
}
\arguments{
\item{config}{API Configuration}
}
\value{
HTTP Status Code
}
\description{
Tests Learning Locker API Configuration
}
\details{
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html
}
